pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Client library for MCP (Model Control Protocol) implementations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.3.0",
    "dotenv>=0.9.9",
    "mcp>=1.1.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.11.10",
    "openai>=1.81.0",
    "streamlit>=1.45.1",
]
